Made function return double instead of int

+++ include/usv_sim_2d/usv.hpp
@@ -9,7 +9,7 @@ class USV
 public:
     struct VechicleState
     {
-        double timestamp = 0;
+        double timestamp = 0.0;
 
         Eigen::Vector3d gyro;
         Eigen::Vector3d accel;
@@ -26,7 +26,7 @@ public:
     bool update(std::array<uint16_t, 16> servo_out);
 
 protected:
-    uint64_t micros();
+    double micros();
 
 private:
     double interval_map(const double &x, const double &x0, const double &x1, const double &y0, const double &y1);
